añadimos la conexión a MySQL y gestionamos logs y fichero.

+++ Docker/main.py
@@ -1,22 +1,37 @@
 def main():
-    ###Variables para check
-    error = 0
-
     ###Creamos el fichero de log
     comprobar.checkErrorsLogs()
 
-    ###Leemos el fichero
-    numeros = leerfichero.LecturaFichero('/opt/files/numeros.txt')
-    ###Comprobamos errores en el fichero
-    error = comprobar.checkErrorFile(numeros)
-    if error > 0:
-        os.system("mv /opt/files/numeros.txt /opt/files/numeros_error.txt")
-        exit()
-    else:
-        sql_numeros = sql_querys.numerosSQL(numeros)
-        # sql_millones = sql_querys.millonesSQL(numeros) !! FALTA SACAR EL ID DEL SORTEO PARA AÑADIR A MILLONES
+    while True:
+        if (os.path.exists('/opt/files/numeros.txt')):
+            ###Leemos el fichero
+            numeros = leerfichero.LecturaFichero('/opt/files/numeros.txt')
+            
+            ###Variables para check
+            error = 0
+            ultimoID = 0
+
+            ###Comprobamos errores en el fichero
+            error = comprobar.checkErrorFile(numeros)
+            if error > 0:
+                os.system("mv /opt/files/numeros.txt /opt/files/numeros_error.txt")
+                exit()
+            else:
+                sql_numeros = sql_querys.numerosSQL(numeros)
+                miConexio = sql_connection.conexionDB()
+                ultimoID = miConexio.ultimoID()
+                sql_millones = sql_querys.millonesSQL(numeros,ultimoID)
+                result1 = miConexio.insertNumeros(sql_numeros)
+                miConexio2 = sql_connection.conexionDB()
+                result2 = miConexio2.insertMillones(sql_millones)
+                if (result1 == 0) and (result2 == 0):
+                    comprobar.successUpdate()
+                    os.system('rm /opt/files/numeros.txt')
+                else:
+                    comprobar.errorUpdate()
 
 if __name__ == '__main__':
+    import sql_connection
     import leerfichero
     import comprobar
     import os
